Allow the use of Faraday 2.x
This allows the use of Faraday 2.x and adds a fix for initializing the
SDK when the config is not set.

Faraday 2 accepts a lambda function for Auth, which actually allows
setting the API Key after initialization. Without the lambda function
and in Faraday 1.x there's actually an issue where the new API key
wouldn't be picked up anyway because the SDK is already initialized.